chore: merge lints into single file
